Create new SafeBufferedAppender
Same principle as BufferedAppender, but fix a potential issue of loosing item !
Here we alway add message to buffer, and then try to flush the buffer.
If item is not sent, we keep reference in memory and try next time.

Could be used in addition with useQueue "autoFlushPeriod" for more convenience.